+++ yandex/data_source_yandex_ydb_database_serverless_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
+
 	"github.com/yandex-cloud/go-genproto/yandex/cloud/ydb/v1"
 )