+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 0.64.0 (Unreleased)
+FEATURES:
+* **New Resource:** `yandex_logging_group`
+* **New Data Source:** `yandex_logging_group`
+
 BUG FIXES:
-* fix sni handler expanding.
-* fix internal_ipv4_address missing subnet_id field
+* fix sni handler expanding
+* fix `internal_ipv4_address` missing `subnet_id` field
 
 ## 0.63.1 (September 10, 2021)
 BUG FIXES: